Renames UILayout to just Layout
