+++ components/App/App.tsx
@@ -6,8 +6,8 @@ import Tooltip, { TooltipContextProvider } from 'components/Tooltip';
 import DetailPanel from 'components/DetailPanel';
 import ControlBar from 'components/ControlBar';
 import { layouts } from 'lib/xbars';
-import Xbar from 'components/UILayout/Xbar';
-import Hotbar from 'components/UILayout/Hotbar';
+import Xbar from '@/components/Layout/Xbar';
+import Hotbar from '@/components/Layout/Hotbar';
 import ActionPanel from 'components/ActionPanel';
 import SystemMessage from 'components/SystemMessage';
 import { SelectedActionContextProvider } from 'components/SelectedAction';
