+++ components/Layout/Hotbar/Hotbar.module.scss
